linkedin_job_scraper/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.3
django-enumfields==2.1.1
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
linkedin_api==2.1.1
lxml==5.1.1
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
